+++ tfhe-benchmark/benches/integer/bench.rs
@@ -151,7 +151,7 @@ fn bench_server_key_binary_function_clean_inputs<F>(
             &benchmark_spec,
             display_name,
             &OperatorType::Atomic,
-            bit_size as u32,
+            bit_size as u64,
             vec![param.message_modulus().0.ilog2(); num_block],
         );
     }
@@ -224,7 +224,7 @@ fn bench_server_key_unary_function_dirty_inputs<F>(
             &benchmark_spec,
             display_name,
             &OperatorType::Atomic,
-            bit_size as u32,
+            bit_size as u64,
             vec![param.message_modulus().0.ilog2(); num_block],
         );
     }
@@ -331,7 +331,7 @@ fn bench_server_key_unary_function_clean_inputs<F>(
             &benchmark_spec,
             display_name,
             &OperatorType::Atomic,
-            bit_size as u32,
+            bit_size as u64,
             vec![param.message_modulus().0.ilog2(); num_block],
         );
     }
@@ -460,7 +460,7 @@ fn bench_server_key_binary_scalar_function_clean_inputs<F, G>(
             &benchmark_spec,
             display_name,
             &OperatorType::Atomic,
-            bit_size as u32,
+            bit_size as u64,
             vec![param.message_modulus().0.ilog2(); num_block],
         );
     }
@@ -605,7 +605,7 @@ fn if_then_else_parallelized(c: &mut Criterion) {
             &benchmark_spec,
             display_name,
             &OperatorType::Atomic,
-            bit_size as u32,
+            bit_size as u64,
             vec![param.message_modulus().0.ilog2(); num_block],
         );
     }
@@ -719,7 +719,7 @@ fn flip_parallelized(c: &mut Criterion) {
             &benchmark_spec,
             display_name,
             &OperatorType::Atomic,
-            bit_size as u32,
+            bit_size as u64,
             vec![param.message_modulus().0.ilog2(); num_block],
         );
     }
@@ -845,7 +845,7 @@ fn ciphertexts_sum_parallelized(c: &mut Criterion) {
                 &benchmark_spec,
                 display_name,
                 &OperatorType::Atomic,
-                bit_size as u32,
+                bit_size as u64,
                 vec![param.message_modulus().0.ilog2(); num_block],
             );
         }
@@ -1374,7 +1374,7 @@ mod cuda {
                 &benchmark_spec,
                 display_name,
                 &OperatorType::Atomic,
-                bit_size as u32,
+                bit_size as u64,
                 vec![param.message_modulus().0.ilog2(); num_block],
             );
         }
@@ -1523,7 +1523,7 @@ mod cuda {
                 &benchmark_spec,
                 display_name,
                 &OperatorType::Atomic,
-                bit_size as u32,
+                bit_size as u64,
                 vec![param.message_modulus().0.ilog2(); num_block],
             );
         }
@@ -1666,7 +1666,7 @@ mod cuda {
                 &benchmark_spec,
                 display_name,
                 &OperatorType::Atomic,
-                bit_size as u32,
+                bit_size as u64,
                 vec![param.message_modulus().0.ilog2(); num_block],
             );
         }
@@ -1822,7 +1822,7 @@ mod cuda {
                 &benchmark_spec,
                 "if_then_else",
                 &OperatorType::Atomic,
-                bit_size as u32,
+                bit_size as u64,
                 vec![param.message_modulus().0.ilog2(); num_block],
             );
         }
@@ -2699,7 +2699,7 @@ mod cuda {
                     &benchmark_spec,
                     display_name,
                     &OperatorType::Atomic,
-                    bit_size as u32,
+                    bit_size as u64,
                     vec![param.message_modulus().0.ilog2(); num_blocks],
                 );
             }
@@ -2942,7 +2942,7 @@ mod hpu {
                 &benchmark_spec,
                 display_name,
                 &OperatorType::Atomic,
-                bit_size as u32,
+                bit_size as u64,
                 vec![param.message_modulus().0.ilog2(); num_block],
             );
         }
@@ -3465,7 +3465,7 @@ fn bench_server_key_cast_function<F>(
                 &benchmark_spec,
                 display_name,
                 &OperatorType::Atomic,
-                bit_size as u32,
+                bit_size as u64,
                 vec![param.message_modulus().0.ilog2(); num_blocks],
             );
         }
